move kprobe::Registers to separate registers module
Registers is shared between kprobes and uprobes, so it's better to have it in a
separate module. Then it's still re-exported by preludes as before.
