+++ redbpf-probes/src/lib.rs
@@ -55,6 +55,7 @@ pub mod helpers;
 pub mod kprobe;
 pub mod maps;
 pub mod net;
+pub mod registers;
 pub mod socket_filter;
 pub mod uprobe;
 pub mod xdp;
